move router files
