+++ server/router/v1/auth/logout/index.ts
