+++ server/router/v1/auth/refresh/index.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 
 import { refreshAccessToken } from '../../../../helper/refreshAccessToken';
-import { TOKEN_EXPIRE_IN } from '../../index';
+import { TOKEN_EXPIRE_IN } from '../../../../config/constants';
 import { SpotifyAPI, ServerAPI } from '~~/types';
 
 type ResponseBody = ServerAPI.Auth.Token
