+++ server/router/v1/index.ts
@@ -11,9 +11,6 @@ dotenv.config();
 
 const router = express.Router();
 
-// 10分でトークンを更新させる
-export const TOKEN_EXPIRE_IN = 1000 * 60 * 10;
-
 router.get('/auth', auth);
 
 router.post('/auth/login', login);
